agente-docente/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
mistralai==1.5.1
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic>=2.5.2
requests==2.32.3 